+++ pool/src/main.rs
@@ -1,13 +1,12 @@
 use bob_miner_v2::{MinerSettings, StatsV2};
 use bob_minter_v2::Stats;
 use bob_pool::guard::GuardPrincipal;
-use bob_pool::guard::TaskGuard;
+use bob_pool::guard::{TaskGuard, TaskType};
 use bob_pool::memory::{
     add_member_total_cycles, add_rewards, commit_block_participants, get_miner_canister,
     get_next_block_participants, get_rewards, set_miner_canister, set_rewards,
     total_pending_rewards,
 };
-use bob_pool::TaskType;
 use bob_pool::{
     fetch_block, notify_top_up, MemberCycles, MAINNET_BOB_CANISTER_ID,
     MAINNET_BOB_LEDGER_CANISTER_ID, MAINNET_CYCLE_MINTER_CANISTER_ID, MAINNET_LEDGER_CANISTER_ID,
